rename @/modules -> @/vuexModules
