+++ src/components/listing/FileLister.vue
@@ -7,10 +7,10 @@ import { Component, Vue, Prop } from 'vue-property-decorator'
 import { NoCache } from 'vue-class-decorator'
 import { MutationPayload } from 'vuex'
 import { getModule } from 'vuex-module-decorators'
-import ListModule from '../../modules/ListModule'
-import Web3Module from '../../modules/Web3Module'
-import MetaMaskModule from '../../modules/MetaMaskModule'
-import FfaListingsModule from '../../modules/FfaListingsModule'
+import ListModule from '../../vuexModules/ListModule'
+import Web3Module from '../../vuexModules/Web3Module'
+import MetaMaskModule from '../../vuexModules/MetaMaskModule'
+import FfaListingsModule from '../../vuexModules/FfaListingsModule'
 import ListingModule from '../../functionModules/protocol/ListingModule'
 import FfaListing from '../../models/FfaListing'
 import { ProcessStatus } from '../../models/ProcessStatus'
