+++ src/components/listing/FileMetadata.vue
@@ -30,11 +30,11 @@ import { Component, Vue, Prop, Watch } from 'vue-property-decorator'
 import { NoCache } from 'vue-class-decorator'
 import { MutationPayload } from 'vuex'
 import { getModule } from 'vuex-module-decorators'
-import UploadModule from '../../modules/UploadModule'
-import FfaListingsModule from '../../modules/FfaListingsModule'
-import TitleFieldValidator from '../../modules/validators/TitleFieldValidator'
-import FfaFieldValidation from '../../modules/validators/FfaFieldValidation'
-import TaggersModule from '../../modules/TaggersModule'
+import UploadModule from '../../vuexModules/UploadModule'
+import FfaListingsModule from '../../vuexModules/FfaListingsModule'
+import TitleFieldValidator from '../../vuexModules/validators/TitleFieldValidator'
+import FfaFieldValidation from '../../vuexModules/validators/FfaFieldValidation'
+import TaggersModule from '../../vuexModules/TaggersModule'
 import { ProcessStatus } from '../../models/ProcessStatus'
 import StartListingButton from '../listing/StartListingButton.vue'
 import TextField from '@/components/ui/TextField.vue'
