+++ src/components/listing/FileUploader.vue
@@ -26,9 +26,9 @@ import { Component, Vue, Prop } from 'vue-property-decorator'
 import { NoCache } from 'vue-class-decorator'
 import { MutationPayload } from 'vuex'
 import { getModule } from 'vuex-module-decorators'
-import UploadModule from '../../modules/UploadModule'
-import UploadModuleValidator from '../../modules/validators/UploadModuleValidator'
-import ListModule from '../../modules/ListModule'
+import UploadModule from '../../vuexModules/UploadModule'
+import UploadModuleValidator from '../../vuexModules/validators/UploadModuleValidator'
+import ListModule from '../../vuexModules/ListModule'
 import FfaListing from '../../models/FfaListing'
 import { ProcessStatus } from '../../models/ProcessStatus'
 import { FileDropped } from '../../models/Events'
