+++ src/functionModules/components/FfaTagModule.ts
@@ -1,4 +1,4 @@
-import TaggersModule from 'modules/TaggersModule'
+import TaggersModule from 'vuexModules/TaggersModule'
 
 export default class FfaTagModule {
   public static deleteTag(taggersModule: TaggersModule, taggerKey: string, tag: string) {
