+++ src/views/List.vue
@@ -21,7 +21,7 @@
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator'
 import { getModule } from 'vuex-module-decorators'
-import FlashesModule from '../modules/FlashesModule'
+import FlashesModule from '../vuexModules/FlashesModule'
 import { FlashType } from '../models/Flash'
 import Flash from '../models/Flash'
 import FlashMessage from '@/components/ui/FlashMessage.vue'
@@ -31,8 +31,8 @@ import FileMetadata from '@/components/listing/FileMetadata.vue'
 import Status from '@/components/ui/Status.vue'
 import Dropzone from 'dropzone'
 import StartListingButton from '../components/listing/StartListingButton.vue'
-import MetaMaskModule from '../modules/MetaMaskModule'
-import Web3Module from '../modules/Web3Module'
+import MetaMaskModule from '../vuexModules/MetaMaskModule'
+import Web3Module from '../vuexModules/Web3Module'
 import MetaMask from '../models/MetaMask'
 
 import '@/assets/style/views/list.sass'
