+++ src/vuexModules/FlashesModule.ts
