+++ src/vuexModules/TaggersModule.ts
