+++ src/vuexModules/VoteModule.ts
