+++ tests/unit/components/listing/FileUploader.spec.ts
@@ -8,10 +8,10 @@ import { library } from '@fortawesome/fontawesome-svg-core'
 import { faFile as faFileSolid } from '@fortawesome/free-solid-svg-icons'
 import { faFile, faCheckCircle } from '@fortawesome/free-regular-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
-import UploadModule from '../../../../src/modules/UploadModule'
-import MetaMaskModule from '../../../../src/modules/MetaMaskModule'
-import Web3Module from '../../../../src/modules/Web3Module'
-import TaggersModule from '../../../../src/modules/TaggersModule'
+import UploadModule from '../../../../src/vuexModules/UploadModule'
+import MetaMaskModule from '../../../../src/vuexModules/MetaMaskModule'
+import Web3Module from '../../../../src/vuexModules/Web3Module'
+import TaggersModule from '../../../../src/vuexModules/TaggersModule'
 import { ProcessStatus, ProcessStatusLabelMap } from '../../../../src/models/ProcessStatus'
 import Web3 from 'web3'
