+++ tests/unit/components/ui/Status.spec.ts
@@ -9,7 +9,7 @@ import { faFile as faFileSolid } from '@fortawesome/free-solid-svg-icons'
 import { faFile, faCheckCircle } from '@fortawesome/free-regular-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import FileUploader from '@/components/listing/FileUploader.vue'
-import UploadModule from '../../../../src/modules/UploadModule'
+import UploadModule from '../../../../src/vuexModules/UploadModule'
 import { ProcessStatus, ProcessStatusLabelMap } from '../../../../src/models/ProcessStatus'
 
 const localVue = createLocalVue()
